🩹 fix : 허브 feignClient 관련 수정

+++ delivery-service/src/main/java/com/logistics/platform/delivery_service/delivery/application/dto/HubRouteResponseDto.java
@@ -15,7 +15,8 @@ public class HubRouteResponseDto {
 
   private UUID startHubId;
   private UUID endHubId;
-  private Long estimatedDuration;
-  private Long estimatedDistance;
+  private Double estimatedDuration;
+  private Double estimatedDistance;
 
 }
+
